requirements.txt
altgraph==0.17.2
appdirs==1.4.4
Babel==2.9.1
beautifulsoup4==4.10.0
boto3==1.20.28
botocore==1.23.28
browser-cookie3==0.12.1
bsdiff4==1.2.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
cryptography==3.4.8
Deprecated==1.2.13
dsdev-utils==1.0.5
idna==3.2
importlib-metadata==4.8.1
jeepney==0.7.1
jmespath==0.10.0
keyring==23.2.1
lz4==3.1.3
macholib==1.15.2
packaging==21.3
pbkdf2==1.3
pbr==5.8.0
Pillow==8.4.0
pyaes==1.6.1
pycparser==2.20
pycryptodome==3.10.1
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
PyNaCl==1.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyUpdater==4.0
PyUpdater-S3-Plugin==4.1.2
requests==2.26.0
s3transfer==0.5.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.2.1
stevedore==3.5.0
text-unidecode==1.3
tkcalendar==1.6.1
tqdm==4.62.3
urllib3==1.26.6
wrapt==1.13.3
zipp==3.5.0
